Test fails on first test run and passed after

+++ spec/exporter_spec.rb
@@ -9,14 +9,14 @@ describe Exporter do
 
   context 'target_path' do
     it 'works with no file separator at end' do
-      Dir.rmdir TARGET_PATH
+      Dir.rmdir TARGET_PATH if File.exist? TARGET_PATH
       exporter.target_path TARGET_PATH
       expect(exporter.target_path).to eq "#{TARGET_PATH}/"
       expect(Dir).to exist(TARGET_PATH)
     end
 
     it 'works with file separator at end' do
-      Dir.rmdir TARGET_PATH
+      Dir.rmdir TARGET_PATH if File.exist? TARGET_PATH
       exporter.target_path "#{TARGET_PATH}/"
       expect(exporter.target_path).to eq "#{TARGET_PATH}/"
       expect(Dir).to exist(TARGET_PATH)
